+++ source/Private/Convert-ShouldBeExactly.ps1
@@ -3,8 +3,8 @@
         Converts a command `Should -BeExactly` to the specified Pester syntax.
 
     .DESCRIPTION
-        The Convert-ShouldBe function is used to convert a command `Should -BeExactly` to
-        the specified Pester syntax.
+        The Convert-ShouldBeExactly function is used to convert a command `Should -BeExactly`
+        to the specified Pester syntax.
 
     .PARAMETER CommandAst
         The CommandAst object representing the command to be converted.
@@ -188,7 +188,12 @@ function Convert-ShouldBeExactly
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }