+++ source/Private/Convert-ShouldBeFalse.ps1
@@ -3,8 +3,8 @@
         Converts a command `Should -BeFalse` to the specified Pester syntax.
 
     .DESCRIPTION
-        The Convert-ShouldBeFalse function is used to convert a command `Should -BeFalse` to
-        the specified Pester syntax.
+        The Convert-ShouldBeFalse function is used to convert a command `Should -BeFalse`
+        to the specified Pester syntax.
 
     .PARAMETER CommandAst
         The CommandAst object representing the command to be converted.
@@ -176,7 +176,12 @@ function Convert-ShouldBeFalse
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }