+++ source/Private/Convert-ShouldBeIn.ps1
@@ -3,8 +3,8 @@
         Converts a command `Should -BeIn` to the specified Pester syntax.
 
     .DESCRIPTION
-        The Convert-ShouldContain function is used to convert a command `Should -BeIn` to
-        the specified Pester syntax.
+        The Convert-ShouldContain function is used to convert a command `Should -BeIn`
+        to the specified Pester syntax.
 
     .PARAMETER CommandAst
         The CommandAst object representing the command to be converted.
@@ -224,7 +224,12 @@ function Convert-ShouldBeIn
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }