+++ source/Private/Convert-ShouldBeLessThan.ps1
@@ -3,8 +3,8 @@
         Converts a command `Should -BeLessThan` to the specified Pester syntax.
 
     .DESCRIPTION
-        The Convert-ShouldBeLessThan function is used to convert a command `Should -BeLessThan` to
-        the specified Pester syntax.
+        The Convert-ShouldBeLessThan function is used to convert a command `Should -BeLessThan`
+        to the specified Pester syntax.
 
     .PARAMETER CommandAst
         The CommandAst object representing the command to be converted.
@@ -193,7 +193,12 @@ function Convert-ShouldBeLessThan
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }