+++ source/Private/Convert-ShouldMatch.ps1
@@ -3,8 +3,8 @@
         Converts a command `Should -Match` to the specified Pester syntax.
 
     .DESCRIPTION
-        The Convert-ShouldMatch function is used to convert a command `Should -Match` to
-        the specified Pester syntax.
+        The Convert-ShouldMatch function is used to convert a command `Should -Match`
+        to the specified Pester syntax.
 
     .PARAMETER CommandAst
         The CommandAst object representing the command to be converted.
@@ -185,7 +185,12 @@ function Convert-ShouldMatch
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }